Refactored ConsoleChannel to DefaultConsoleChannel and added input handling
Refactored the use of ConsoleChannel to DefaultConsoleChannel to make it more expressive about its default nature. Also, enhanced the functionality of DefaultConsoleChannel by adding a method to handle inputs. Renamed the 'consoleChannel' to 'defaultConsoleChannel' at various locations for the sake of consistency. The added functionality makes it possible to handle different types of keys appropriately, enhancing the overall usability of the console.